% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_additional_info.R
\name{additional_info}
\alias{additional_info}
\title{Get additional information from the yaml file about PsN version, Nonmem version, R version and loaded R packages that can be added to the pdf report.}
\usage{
additional_info(yaml_file)
}
\arguments{
\item{yaml_file}{Yaml file name}
}
\value{
A list of three character elements (PsN_version, NONMEM_version and R_version) and one dataframe element (R_packages).
}
\description{
Get additional information from the yaml file about PsN version, Nonmem version, R version and loaded R packages that can be added to the pdf report.
}
