% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_all_structural_extra_plots.R
\name{all_structural_extra_plots}
\alias{all_structural_extra_plots}
\title{Print all structural part plots and tables.}
\usage{
all_structural_extra_plots(simeval_directory, base_dataset,
  resmod_structural_details, extra_table, idv_all, idv_name, dvid_name,
  type, nonlinear)
}
\arguments{
\item{simeval_directory}{A path to the simeval run directory.}

\item{base_dataset}{A path of the linbase.dta (if nonlinear=TRUE) or to the original input data file (if nonlinear=FALSE).}

\item{resmod_structural_details}{}

\item{extra_table}{A path of the extra table with all derivatives columns.}

\item{idv_all}{A string vector of all possible idv names for specific run.}

\item{idv_name}{The string of the idv name from the -idv option in the qa run.}

\item{dvid_name}{A string of the DVID column.}

\item{type}{A string of the Rmarkdown format. Possible values "latex" or "html". Will be passed to the kable() function.}

\item{nonlinear}{A logical indicating whether nonlinear qa have been run.}
}
\value{
Print kable tables and plots to the report.
}
\description{
Print all structural part plots and tables.
}
