% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_break_text.R
\name{break_text}
\alias{break_text}
\title{Break command text in reasonable places so it can fit in a pdf report.}
\usage{
break_text(text, max_symb = 100)
}
\arguments{
\item{text}{A sting}

\item{max_symb}{A number of elements per row. By default max_symb=100.}
}
\value{
A string
}
\description{
Break command text in reasonable places so it can fit in a pdf report.
}
