% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_captions_structural.R
\name{captions_structural}
\alias{captions_structural}
\title{Captions for the structural part plots and tables based on the idv name.}
\usage{
captions_structural(idv_all, idv_name, which_idv, perc)
}
\arguments{
\item{idv_all}{A string vector of all possible idv names for specific run.}

\item{idv_name}{The string of the idv name from the -idv option in the qa run.}

\item{which_idv}{The string of the idv name which captions are needed.}

\item{perc}{A logical if values in the structural tables and plots are presented in percentages.}
}
\value{
A list of the captions for structural plots and tables of specific idv.
}
\description{
Captions for the structural part plots and tables based on the idv name.
}
