% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_check_package_versions.R
\name{check_package_versions}
\alias{check_package_versions}
\title{Check package versions. Will print an error message if some package version is too low.}
\usage{
check_package_versions(pkg, versions, toolname, do_quit = T)
}
\arguments{
\item{pkg}{A character vector with package names}

\item{versions}{A character vector with package versions}

\item{toolname}{A string of the tool name}

\item{do_quit}{A logical indicating whether function should quit if some package version is too low. By default do_quit=TRUE}
}
\description{
Check package versions. Will print an error message if some package version is too low.
}
