% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_command_start.finish.time.R
\name{command_start.finish.time}
\alias{command_start.finish.time}
\title{Get command line text, start and finish time of the run}
\usage{
command_start.finish.time(yaml_file)
}
\arguments{
\item{yaml_file}{A yaml file name}
}
\value{
A list of three elements: command, run_start, run_finish
}
\description{
Get command line text, start and finish time of the run
}
