% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_conditions.R
\name{do_safely}
\alias{do_safely}
\title{Run expression and retain errors}
\usage{
do_safely(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
A list with entries results and error (one of them is NA)
}
\description{
The function allows to evaluates an expression and store eventual errors in the results data structure.
}
\examples{
do_safely("a"/1)
}
