% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_ext_ofv.R
\name{.get_ext_ofv}
\alias{.get_ext_ofv}
\title{Gets objective function value from the specific iteration.}
\usage{
.get_ext_ofv(ext_file, iteration = -1e+09)
}
\arguments{
\item{ext_file}{A string of the ext file name.}

\item{iteration}{A numeric value from the ITERATION column in the ext file. By default iteration=-1000000000.}
}
\value{
An objective function value from specific iteration.
}
\description{
Gets objective function value from the specific iteration.
}
