% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_rawres_ofv.R
\name{.get_rawres_ofv}
\alias{.get_rawres_ofv}
\title{Get a ofv value from the csv file column ofv.}
\usage{
.get_rawres_ofv(rawres_file, row = 1)
}
\arguments{
\item{rawres_file}{A csv file.}

\item{row}{Number of row from which ofv value should be selected. By default row = 1.}
}
\value{
Objective function value.
}
\description{
Get a ofv value from the csv file column ofv.
}
