% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_error_table.R
\name{error_table}
\alias{error_table}
\title{Creates a data frame with ERROR records}
\usage{
error_table(first_column = "", col = 2, column_names)
}
\arguments{
\item{first_column}{A string of the first column name in the output data frame. By default is empty string.}

\item{col}{Number of needed columns of the output data frame. By default is set to 2.}

\item{column_names}{A string vector of the column names. If missing and number of columns are less than 2 then will be set to NULL, 
otherwise second column will be set to "dOFV" and all other columns will be set to empty strings.}
}
\value{
A data frame with the ERROR records.
}
\description{
Creates a data frame with ERROR records
}
