% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_reading.R
\name{ext_file}
\alias{ext_file}
\alias{phi_file}
\title{Return name of .ext or .phi file}
\usage{
ext_file(path)

phi_file(path)
}
\arguments{
\item{path}{Model or lst file for which to return the file}
}
\value{
character
}
\description{
Return name of .ext or .phi file
}
\examples{
ext_file("run4.mod")
phi_file("run4.lst")
}
