% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_add_etas_table.R
\name{get_add_etas_table}
\alias{get_add_etas_table}
\title{Consequences of including additional subject-level random effects on selected parameters in terms of estimated standard deviation (SD) of new
and existing random effect parameters as well as expected improvement in OFV.}
\usage{
get_add_etas_table(
  original_max0_model,
  add_etas_dir,
  added_etas,
  dofv_add.etas,
  nonlinear,
  quiet = F
)
}
\arguments{
\item{original_max0_model}{The base model file name.}

\item{add_etas_dir}{A name of the directory where add_etas_linbase.ext (nonlinear=FALSE) or add_etas.ext (nonlinear=TRUE) can be found.}

\item{added_etas}{A list of additional etas and their numeric position in the model. (added_etas <- list(TPEN=6, SIND=5))}

\item{dofv_add.etas}{Difference between base model and the model with additional etas OFV values.}

\item{nonlinear}{A logical indicating whether nonlinear qa have been run.}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A list of 2 arguments:
add_etas_table - a data frame
add_etas_error - a logical indicating whether both ext files original_max0_model and add_etas_linbase.ext/add_etas.ext exist in the folder
}
\description{
Consequences of including additional subject-level random effects on selected parameters in terms of estimated standard deviation (SD) of new
and existing random effect parameters as well as expected improvement in OFV.
}
