% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_eta.etat_values.R
\name{get_eta.etat_values}
\alias{get_eta.etat_values}
\title{Calculate ETAT values for each eta in the phi file of the parameter model based on the formula: 
If have boxcox model: (exp(eta)^lambda -1)/lambda
If have tdist model: eta(1+((eta^2 + 1)/(4deg_of_freedom))+((5eta^4 + 16eta^2 + 3)/(96deg_of_freedom^2))+((3eta^6 + 19eta^4 + 17eta^2 - 15)/(384deg_of_freedom^3)))}
\usage{
get_eta.etat_values(param_model, theta_values, quiet = F)
}
\arguments{
\item{param_model}{The parameter model file name. Phi file will be used.}

\item{theta_values}{A data frame of lambda values for each ETA (if boxcox model) or a data frame of degrees of freedom for each ETA (if tdist model).}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A data frame
}
\description{
Calculate ETAT values for each eta in the phi file of the parameter model based on the formula: 
If have boxcox model: (exp(eta)^lambda -1)/lambda
If have tdist model: eta(1+((eta^2 + 1)/(4deg_of_freedom))+((5eta^4 + 16eta^2 + 3)/(96deg_of_freedom^2))+((3eta^6 + 19eta^4 + 17eta^2 - 15)/(384deg_of_freedom^3)))
}
