% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_eta_transf_table.R
\name{get_eta_transf_table}
\alias{get_eta_transf_table}
\title{Calculate ETAT values for each eta in normal distribution. Calculations done based on the formula: 
If boxcox model: (exp(eta)^lambda -1)/lambda
If tdist model: eta(1+((eta^2 + 1)/(4deg_of_freedom))+((5eta^4 + 16eta^2 + 3)/(96deg_of_freedom^2))+((3eta^6 + 19eta^4 + 17eta^2 - 15)/(384deg_of_freedom^3)))}
\usage{
get_eta_transf_table(input_table, seq_length.out = 1000)
}
\arguments{
\item{input_table}{A data frame of ETA name in the column ETA, lambda/degrees of freedom values in the column Lambda/Degrees of freedom and 
standard deviation (from parameter model) of the ETA in the third column.}

\item{seq_length.out}{Desired length of the sequence for the seq function. So many eta values for each ETA and each ETAT will be in the output dataframe. 
By default seq_length.out=1000.}
}
\value{
A list of 3 elements:
eta_transf_table - a data frame of the transformed ETA values
fig_height - hight of the figure which is based on how many ETAs are in the table
make_eta_transf_plot - a logical identifier if eta transformation plot should be created
}
\description{
Calculate ETAT values for each eta in normal distribution. Calculations done based on the formula: 
If boxcox model: (exp(eta)^lambda -1)/lambda
If tdist model: eta(1+((eta^2 + 1)/(4deg_of_freedom))+((5eta^4 + 16eta^2 + 3)/(96deg_of_freedom^2))+((3eta^6 + 19eta^4 + 17eta^2 - 15)/(384deg_of_freedom^3)))
}
