% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ext_file_manipulations.R
\name{get_ext_ofv}
\alias{get_ext_ofv}
\title{Get ofv value from specific iteration row}
\usage{
get_ext_ofv(filename, iteration = -1e+09)
}
\arguments{
\item{filename}{Name of the file.}

\item{iteration}{The value of iteration. By default is -1000000000.}
}
\value{
The OFV value.
}
\description{
Get ofv value from specific iteration row
}
