% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ext.R
\name{get_final_ofvs}
\alias{get_final_ofvs}
\title{Get final OFVs}
\usage{
get_final_ofvs(ext_list, table_selection = 1)
}
\arguments{
\item{ext_list}{A data structure as returned by \code{\link{read_nm_ext}}}

\item{table_selection}{A vector to select the list entries in ext_list}
}
\value{
A vector of OFV values
}
\description{
Get final OFVs
}
