% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_ii_table.R
\name{get_ii_table}
\alias{get_ii_table}
\title{Get influential individuals table.}
\usage{
get_ii_table(
  cdd_directory,
  model.filename,
  cutoff,
  max_rows,
  skip = NULL,
  nonlinear,
  quiet = F
)
}
\arguments{
\item{cdd_directory}{Cdd run directory. Will search for the files skipped_individuals1.csv and
raw_results_"model name"_linbase.csv (if nonlinear=F) or raw_results_"model name".csv (if nonlinear=T)}

\item{model.filename}{The base model file name.}

\item{cutoff}{The dOFV cutoff value for influential individual identification.}

\item{max_rows}{Number of infuential individuals to print in the pdf file.}

\item{skip}{A character vector with names of the skipped parts in the qa run. Will check if "cdd" is one of the vector elements.
By default skip=NULL.}

\item{nonlinear}{A logical indicating whether nonlinear qa have been run.}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A list of 7 arguments.
cdd_files_exist - a logical indicating whether needed files exist in the input directory.
cdd.data - a data frame of all id numbers and dOFV values for each id.
all_dofv - a vector of all dofv values.
cdd_highest_dofv - A data frame of one most influential individual. Will be used for summary table.
ii_table - A data frame of all influential individuals, where max row number is set to max_rows.
infl_id - A numeric vector with all influential individual ID numbers.
fig_height_infl - A numeric for idividual plots hight.
}
\description{
Get influential individuals table.
}
