% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ext_file_manipulations.R
\name{get_initial_estimate_values}
\alias{get_initial_estimate_values}
\title{Get initial estimate values from specifick iteration row.}
\usage{
get_initial_estimate_values(
  filename,
  select = "all",
  iteration = -1e+09,
  do.stop = TRUE
)
}
\arguments{
\item{filename}{Name of the file.}

\item{select}{Can be set to "omega","theta","sigma" or "all"(means thetas, omegas and sigmas). By default is "all".}

\item{iteration}{The value of iteration. By default is -1000000000.}

\item{do.stop}{A logical indicating whether function should stop if argument 'select' is set to something different than allowed string. By default do.stop=TRUE}
}
\value{
The data frame of initial estimate values.
}
\description{
Get initial estimate values from specifick iteration row.
}
