% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_phi.R
\name{get_iofv_sum}
\alias{get_iofv_sum}
\title{Get sum of iOFV}
\usage{
get_iofv_sum(phi_list, table_selection = 1)
}
\arguments{
\item{phi_list}{A data structure as returned by \code{\link{read_nm_phi}}}

\item{table_selection}{A vector to select the list entries in phi_list}
}
\value{
A numeric vector of length table_selection
}
\description{
Get sum of iOFV
}
