% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_iov_table.R
\name{get_iov_table}
\alias{get_iov_table}
\title{Consequences of including inter-occasion variability in terms of estimated standard deviation (SD) for inter-individual (IIV)
and inter-occasion variability (IOV) as well as expected improvement in OFV.}
\usage{
get_iov_table(original_max0_model, iov_model, iov_etas, dofv_iov, quiet = F)
}
\arguments{
\item{original_max0_model}{The base model file name.}

\item{iov_model}{The iov model file name.}

\item{iov_etas}{A list of occasion vectors of eta numbers. (iov_etas <- list(occ1=c(1, 2, 5, 7), occ2=c(3, 4, 6, 8)))}

\item{dofv_iov}{Difference between base model and the model with added iov OFV values.}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A list of 2 elements:
iov_table - a data frame
iov_error - a logical indicating whether both ext files from models original_max0_model and iov_model exist in the folder
}
\description{
Consequences of including inter-occasion variability in terms of estimated standard deviation (SD) for inter-individual (IIV)
and inter-occasion variability (IOV) as well as expected improvement in OFV.
}
