% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_read_results.R
\name{get_ofv_summary}
\alias{get_ofv_summary}
\title{Retrieve OFV summary table}
\usage{
get_ofv_summary(nonlin_ext_path, lin_ext_path, lin_phi_path)
}
\arguments{
\item{nonlin_ext_path}{Path to derivatives.ext}

\item{lin_ext_path}{Path to linbase.ext}

\item{lin_phi_path}{Path to linbase.phi}
}
\value{
data.frame(name,ofv,result)
}
\description{
Retrieve OFV summary table
}
