% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_outliers_table.R
\name{get_outliers_table}
\alias{get_outliers_table}
\title{Get outliers table.}
\usage{
get_outliers_table(simeval_directory, cdd.data, skip = NULL, quiet = F)
}
\arguments{
\item{simeval_directory}{Simeval run directory. Will search for the files raw_all_iofv.csv}

\item{cdd.data}{a data frame of all id numbers and dOFV values for each id from the cdd run.}

\item{skip}{A character vector with names of the skipped parts in the qa run. Will check if "simeval" is one of the vector elements.
By default skip=NULL.}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A list of 5 arguments.
simeval_files_exist - a logical indicating whether needed files exist in the input directory.
outliers_table - A data frame of all outlying individuals. Max 10 rows.
max_outlier_table - A data frame of one most outlying individual. Will be used for summary table.
outlier_ids - A numeric vector with all outlying individual ID numbers.
fig_height_outl - A numeric for idividual plots hight.
}
\description{
Get outliers table.
}
