% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_overview_table.R
\name{get_overview_table}
\alias{get_overview_table}
\title{Get overview table.}
\usage{
get_overview_table(structural_overview, param_var_overview,
  covariates_overview, resmod_ruv_overview, infl_indiv_overview,
  outliers_overview)
}
\arguments{
\item{structural_overview}{a data frame of structural models, corresponding dofv values and number of added parameters}

\item{param_var_overview}{a data frame of patameter models, corresponding dofv values and number of added parameters}

\item{covariates_overview}{a data frame of covariate models, corresponding dofv values and number of added parameters}

\item{resmod_ruv_overview}{a data frame of residual error models, corresponding dofv values and number of added parameters}

\item{infl_indiv_overview}{a data frame of most influential individual, corresponding dofv value and number of added parameters}

\item{outliers_overview}{a data frame of most outlying individual, corresponding dofv value and number of added parameters}
}
\value{
A list of 4 arguments.
overview_table - a data frame
n.rgroup - a numeric vector with number of rows in each group
rgroup_names - a character vector of group names
row_groups - a data frame with group names as well as a cutoff for the first and last row number in the overview_table for each group
}
\description{
Get overview table.
}
