% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_param_extra_table.R
\name{get_param_extra_table}
\alias{get_param_extra_table}
\title{Get table of estimated shape parameters (Lambda)/degrees of freedom, estimated standard deviation (sd) values 
as well as expected improvement in OFV when estimating a Box-Cox/t-distribution transformation for each subject-level random effect.}
\usage{
get_param_extra_table(original_max0_model, param_model, dofv, quiet = F)
}
\arguments{
\item{original_max0_model}{The base model file name.}

\item{param_model}{The parameter model (box-cox or tdist) file name.}

\item{dofv}{Difference between base and the boxcox or tdist transformed model OFV values.}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A list of 3 elements:
param_extra_table - a data frame
param_extra_table_orig - a data frame
param_extra_table_error - a logical indicating whether both input files original_max0_model and param_model exist in the folder

Difference between param_extra_table and param_extra_table_orig data frames are that values are not formated or rounded in the param_extra_table_orig and 
the dOFV value is not added to this data frame.
}
\description{
Get table of estimated shape parameters (Lambda)/degrees of freedom, estimated standard deviation (sd) values 
as well as expected improvement in OFV when estimating a Box-Cox/t-distribution transformation for each subject-level random effect.
}
