% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_param_var_tables.R
\name{get_param_var_tables}
\alias{get_param_var_tables}
\title{Summary of the transformed models dOFV values and added parameters.}
\usage{
get_param_var_tables(directory, base_model, skip = NULL, quiet = F)
}
\arguments{
\item{directory}{A string of the directory name where modelfit_run, add_etas_run folders and base model ext file can be found. 
In modelfit_run folder expecting to have mod and ext files from fullblock, boxcox, tdist and iov runs. 
In add_etas_run expecting to have mod and ext files from the add_etas_linbase.}

\item{base_model}{A string of the model file name that was transformed.}

\item{skip}{A character vector with names of the skipped parts in the qa run. Will check if "transform" is one of the vector elements.
By default skip=NULL.}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A list of 11 elements:
par_var_models - a data frame with dOFV values of the transformed models and number of added parameters to each of the transformed models
dofv_block - dofv values of full omega block transformed model
dofv_box - dofv values of box-cox transformed model
dofv_tdist - dofv values of t-dist transformed model
dofv_add_etas - dofv values of transformed model with added etas
dofv_iov - dofv values of transformed model with added iov
fullblock_mod - a logical indicating whether transformed model (fullblock.mod) exist in the folder
boxcox_mod - a logical indicating whether transformed model (boxcox.mod) exist in the folder
add_etas_mod - a logical indicating whether transformed model (add_etas_linbase.mod) exist in the folder
tdist_mod - a logical indicating whether transformed model (tdist.mod) exist in the folder
iov_mod - a logical indicating whether transformed model (iov.mod) exist in the folder

In case if string "transform" is one of the 'skip' vector elements, all dOFV values of the output data frame will be replaced with the string "SKIPPED".
In case of missing ext file of the base model all dOFV values will be replaced with the string "ERROR".
In case of missing transformation model file corresponding dOFV value will be replaced with the string "NA".
In case of missing ext file of the transformation model corresponding dOFV value will be replaced with the string "ERROR".
}
\description{
Summary of the transformed models dOFV values and added parameters.
}
