% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_csv_file_manipulation.R
\name{get_rawres_ofv}
\alias{get_rawres_ofv}
\title{Get a ofv value from the csv file.}
\usage{
get_rawres_ofv(filename, row = 1)
}
\arguments{
\item{filename}{Name of the file.}

\item{row}{The row number from which ofv value will be taken. By default is 1.}
}
\value{
A ofv value of specific row.
}
\description{
Get a ofv value from the csv file.
}
