% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_get_scm_table.R
\name{get_scm_table}
\alias{get_scm_table}
\title{Table of model parameter-covariate relationship, resulting improvement in OFV and estimated covariate coefficient.}
\usage{
get_scm_table(scm_directory, nr_cov, nr_param, skip = NULL, quiet = F)
}
\arguments{
\item{scm_directory}{Scm run directory. Will search for the file raw_results_scm.csv. Columns relation, ofv, step.number 
and all columns with estimated covariate coefficient values (like, CLWGT.4.1) will be used in the function.}

\item{nr_cov}{Number of covariates from the scm point of view, because categorical covariates can be binarized.}

\item{nr_param}{Numbner of parameters.}

\item{skip}{A character vector with names of the skipped parts in the qa run. Will check if "scm" is one of the vector elements.
By default skip=NULL.}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A list of 3 elements:
scm_table - a data frame of model parameter covariate relationship, resulting improvement in OFV and estimated covariate coefficient
max_scm_table - a data frame with only one row from scm_table with the highest dofv value
scm_files_exists - logical argument indicating whether needed file exists in the input directory and whether there is at least one covariate and one parameter.
}
\description{
Table of model parameter-covariate relationship, resulting improvement in OFV and estimated covariate coefficient.
}
