% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_hist_of_cdd_dofv.R
\name{hist_of_cdd_dofv}
\alias{hist_of_cdd_dofv}
\title{A histogram of distribution of OFV improvements (dOFV) when excluding specific subjects during the fit.}
\usage{
hist_of_cdd_dofv(values, quiet = F)
}
\arguments{
\item{values}{All cdd dofv values. Missing values will be excluded.}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A ggplot object.
}
\description{
A histogram of distribution of OFV improvements (dOFV) when excluding specific subjects during the fit.
}
