% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frem_ID_ratio_function.R
\name{ID_ratio}
\alias{ID_ratio}
\title{Plots of most extreme individuals (in both directions) in terms of expected effect size of their covariate sets, on parameter (assuming log-normal). 
The bands show 90% confidence interval around the point estimate.}
\usage{
ID_ratio(frem_id, covdata, pardata)
}
\arguments{
\item{frem_id}{- a data frame}

\item{covdata}{- covariate data frame with columns covname, perc5th, mean, perc95th, reference, is.categorical, unit, category.reference and category.other}

\item{pardata}{- parameter data frame with column parname}
}
\value{
Function does all the calculations and the creates a list of plots.
}
\description{
Plots of most extreme individuals (in both directions) in terms of expected effect size of their covariate sets, on parameter (assuming log-normal). 
The bands show 90% confidence interval around the point estimate.
}
