% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_individual_plots_extra.R
\name{individual_plots_extra}
\alias{individual_plots_extra}
\title{Observations (DV), individual predictions (IPRED) and population predictions (PRED) versus time for subjects identified as influential.}
\usage{
individual_plots_extra(file_name, ID_nr, dvid_name, quiet = F)
}
\arguments{
\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}

\item{values}{All cdd dofv values. Missing values will be excluded.}
}
\value{
A ggplot object.
}
\description{
Observations (DV), individual predictions (IPRED) and population predictions (PRED) versus time for subjects identified as influential.
}
