% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_iofv.R
\name{iofv_vs_iofv}
\alias{iofv_vs_iofv}
\title{Agreement in individual OFV contribution (iOFV) plot between between two models.}
\usage{
iofv_vs_iofv(phi1, phi2, quiet = F)
}
\arguments{
\item{phi1}{A string of the first model phi file name.}

\item{phi2}{A string of the second model phi file name.}

\item{quiet}{A logical indicating whether function should not write the warning message if some file not found. By default quiet=FALSE.}
}
\value{
A list of two elements:
make_plot - logical argument indicating whether needed files exist.
plot - a ggplot plot object, scatter plot.
}
\description{
Agreement in individual OFV contribution (iOFV) plot between between two models.
}
