% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_kable_table.R
\name{kable_table}
\alias{kable_table}
\title{Create a kable table basen on chosen Rmarkdown format (latex or html). Special case if only one column in the input dataframe.}
\usage{
kable_table(table, format, ...)
}
\arguments{
\item{table}{A data frame.}

\item{format}{A string of the Rmarkdown format. Possible values "latex" or "html". Not tested with other values like "markdown", "pandoc", and "rst".}

\item{...}{All other arguments that can be passed to the kable() function.}
}
\value{
A kable dataframe.
}
\description{
Create a kable table basen on chosen Rmarkdown format (latex or html). Special case if only one column in the input dataframe.
}
