% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_meta.R
\name{model_prefix}
\alias{model_prefix}
\title{The prefix is what comes before the run number and the extension
i.e. 'run' for 'run1.mod'}
\usage{
model_prefix(meta)
}
\description{
The prefix is what comes before the run number and the extension
i.e. 'run' for 'run1.mod'
}
