% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frem_parameter_ratio_function.R
\name{parameter_ratio}
\alias{parameter_ratio}
\title{Plots of 5th and 95th percentile of each covariate and the size of the effect on parameter (assuming log-normal).
The bands show 90\% confidence interval around the point estimate. Categorical covariates are compared to the other category.}
\usage{
parameter_ratio(inTable_frem, covdata, pardata)
}
\arguments{
\item{inTable_frem}{- ratio data frame of 5th and 95th percentile of each covariate and the size of the effect on parameter}

\item{covdata}{- covariate data frame with columns covname, perc5th, mean, perc95th, reference, is.categorical, unit, category.reference and category.other}

\item{pardata}{- parameter data frame with column parname}
}
\value{
Function does all the calculations and the creates a list of plots.
}
\description{
Plots of 5th and 95th percentile of each covariate and the size of the effect on parameter (assuming log-normal).
The bands show 90\% confidence interval around the point estimate. Categorical covariates are compared to the other category.
}
