% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_plot.R
\name{prepare_parvar_vpc}
\alias{prepare_parvar_vpc}
\title{Prepare VPC for parvar results}
\usage{
prepare_parvar_vpc(
  qa_results,
  transformation = c("fullblock", "boxcox", "tdist")
)
}
\arguments{
\item{qa_results}{QA result data structure}

\item{transformation}{Name of the parvar transformation to create the VPC for}
}
\value{
A result data structure
}
\description{
Prepare VPC for parvar results
}
