% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_plot.R
\name{prepare_va_cov_plot}
\alias{prepare_va_cov_plot}
\title{Prepare VA plot of FREM results}
\usage{
prepare_va_cov_plot(qa_results)
}
\arguments{
\item{qa_results}{QA result data structure}
}
\value{
A result data structure
}
\description{
Prepare VA plot of FREM results
}
