% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsNR-package.R
\docType{package}
\name{PsNR-package}
\alias{PsNR}
\alias{PsNR-package}
\title{PsNR: R package for pharmacometric utilities}
\description{
d
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UUPharmacometrics/PsNR}
}

}
\author{
\strong{Maintainer}: Svetlana Freiberga \email{svetlana.freiberga@farmbio.uu.se} [copyright holder]

Authors:
\itemize{
  \item Rikard Nordgren \email{rikard.nordgren@farmbio.uu.se} [copyright holder]
  \item Sebastian Ueckert \email{sebastian.ueckert@farmbio.uu.se} [copyright holder]
}

}
\keyword{internal}
