% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_html_report.R
\name{qa_html_report}
\alias{qa_html_report}
\title{Render HTML version of QA report}
\usage{
qa_html_report(...)
}
\arguments{
\item{...}{Arguments passed down}
}
\value{
R Markdown output format to pass to render
}
\description{
Render HTML version of QA report
}
