% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_report_content.R
\name{qa_report_content}
\alias{qa_report_content}
\alias{qa_report_content_intro}
\alias{qa_report_content_overview}
\alias{qa_report_content_structural}
\alias{qa_report_content_parvar}
\alias{qa_report_content_full_block}
\alias{qa_report_content_box_cox}
\alias{qa_report_content_tdist}
\alias{qa_report_content_iov}
\alias{qa_report_content_add_eta}
\alias{qa_report_content_residual}
\alias{qa_report_content_influential}
\alias{qa_report_content_outliers}
\alias{qa_report_content_covariates}
\alias{qa_report_content_vaplot}
\alias{qa_report_content_frem_plots}
\title{Content sections for the QA report}
\usage{
qa_report_content_intro()

qa_report_content_overview()

qa_report_content_structural()

qa_report_content_parvar()

qa_report_content_full_block()

qa_report_content_box_cox()

qa_report_content_tdist()

qa_report_content_iov()

qa_report_content_add_eta()

qa_report_content_residual()

qa_report_content_influential()

qa_report_content_outliers()

qa_report_content_covariates()

qa_report_content_vaplot()

qa_report_content_frem_plots()
}
\value{
Latex/HTML for the section
}
\description{
These function generate the different content section for the QA report
}
