% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_reading.R
\name{read_nm_std_sim_table}
\alias{read_nm_std_sim_table}
\title{Read NONMEM simulation table in standard format}
\usage{
read_nm_std_sim_table(path)
}
\arguments{
\item{path}{Path for the table file to read}
}
\value{
A data.frame
}
\description{
This function reads $TABLE output from a simulation step (with possibly repeating headers). This function can
can in general be used instead of \code{\link{read_nm_std_table}} but will be slightly slower due to the
processing of the intermediate header rows.
}
