% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_reading.R
\name{read_nm_std_table}
\alias{read_nm_std_table}
\title{Read NONMEM table in standard format}
\usage{
read_nm_std_table(path, skip = 1)
}
\arguments{
\item{path}{Path for the table file to read}

\item{skip}{Number of lines to skip before the header}
}
\value{
A data.frame
}
\description{
This function reads a NONMEM output table that was generated with the default format option. 
The function should only be used for PsN generated tables with known format and not for "user defined" tables
that might have particular format options.
}
