% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_render.R
\name{render_qa_table}
\alias{render_qa_table}
\alias{render_ofv_table}
\alias{render_covariates_table}
\title{Render table}
\usage{
render_ofv_table(qa_results, settings = qa_settings())

render_covariates_table(qa_results, settings = qa_settings())
}
\arguments{
\item{qa_results}{QA result object as returned by \code{\link{retrieve_qa_results}}}

\item{settings}{Settings object}
}
\value{
Knitted table
}
\description{
Render table
}
\section{Functions}{
\itemize{
\item \code{render_ofv_table}: Renders the OFV overview table in the beginning of the report

\item \code{render_covariates_table}: Renders the covariates table
}}

