% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_read_results.R
\name{retrieve_scm_results}
\alias{retrieve_scm_results}
\title{Retrieve SCM results}
\usage{
retrieve_scm_results(scm_path, parameters, continuous = NULL,
  categorical = NULL)
}
\arguments{
\item{scm_path}{Path to raw_result_scm.csv}

\item{parameters}{Character vector of parameter names}

\item{continuous}{Character vector of continuous covariate names}

\item{categorical}{Character vector of categorical covariate names}
}
\value{
data.frame()
}
\description{
Retrieve SCM results
}
