% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frem_sd_of_unexpl_var_function.R
\name{sd_unexpl_var}
\alias{sd_unexpl_var}
\title{Plots of amount of unexplained variability (on standard deviation scale) remaining in the parameters after knowledge of no covariates, 
each covariate separately and all covariates simultaneously.}
\usage{
sd_unexpl_var(sd_coef_summary, covdata, pardata)
}
\arguments{
\item{sd_coef_summary}{- a data frame}

\item{covdata}{- covariate data frame with columns covname, perc5th, mean, perc95th, reference, is.categorical, unit, category.reference and category.other}

\item{pardata}{- parameter data frame with column parname}
}
\value{
Function does all the calculations and the creates a list of plots.
}
\description{
Plots of amount of unexplained variability (on standard deviation scale) remaining in the parameters after knowledge of no covariates, 
each covariate separately and all covariates simultaneously.
}
