% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_which_resmod_folders.R
\name{which_resmod_folders}
\alias{which_resmod_folders}
\title{Get all idv names. Idv name will be collected from the resmod_'idv' folder name.}
\usage{
which_resmod_folders(directory, idv_name, quiet = F)
}
\arguments{
\item{directory}{The qa run directory path.}

\item{idv_name}{The string of the idv name from the -idv option in the qa run.}

\item{quiet}{A logical indicating whether function should not write the warning message 
if some file not found. By default quiet=FALSE.}
}
\value{
Character vector with all idv names.
}
\description{
Get all idv names. Idv name will be collected from the resmod_'idv' folder name.
}
