% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_results_summary.R
\name{yaml_summary}
\alias{yaml_summary}
\title{Creates a yaml file with all results of the qa run.}
\usage{
yaml_summary(
  model.filename,
  ofv_table,
  resmod_structural_details,
  full_omega_block_list,
  boxcox_lambdas_list,
  tdist_list,
  add_etas_list,
  iov_list,
  frem_table_list,
  scm_table_list,
  resmod_table_list,
  ii_list,
  outlier_table_list,
  overview_table_list,
  dvid_name,
  skip,
  idv_all
)
}
\arguments{
\item{model.filename}{The original model file name.}

\item{ofv_table}{The ofv dataframe from the qa_ofv_summary_table.R function.}

\item{resmod_structural_details}{The output list from the qa_get_resmod_structural_details.R function.}

\item{full_omega_block_list}{The output list from the qa_get_full_omega_block.R function.}

\item{boxcox_lambdas_list}{The output list from the qa_get_param_extra_table.R function for the boxcox model.}

\item{tdist_list}{The output list from the get_param_extra_table.R function for the tdist model.}

\item{add_etas_list}{The output list from the qa_get_add_etas_table.R function.}

\item{iov_list}{The output list from the qa_get_iov_table.R function.}

\item{frem_table_list}{The output list from the qa_get_all_covariates.R function.}

\item{scm_table_list}{The output list from the qa_get_scm_table.R function.}

\item{resmod_table_list}{The output list from the qa_get_resmod_ruv_table.R function.}

\item{ii_list}{The output list from the qa_get_ii_table.R function.}

\item{outlier_table_list}{The output list from the qa_get_outliers_table.R function.}

\item{overview_table_list}{The output list from the qa_get_overview_table.R function.}

\item{dvid_name}{A charaster of the DVID column name.}

\item{skip}{A character vector with names of the skipped parts in the qa run.}

\item{idv_all}{A character vector with all idv names.}
}
\value{
A yaml file results_summary.yaml for the specific qa run.
}
\description{
Creates a yaml file with all results of the qa run.
}
