% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_report_content.R
\name{qa_report_content}
\alias{qa_report_content}
\alias{qa_report_content_intro}
\alias{qa_report_content_overview}
\alias{qa_report_content_structural}
\alias{qa_report_content_parvar}
\alias{qa_report_content_covariates}
\title{Content sections for the QA report}
\usage{
qa_report_content_intro()

qa_report_content_overview()

qa_report_content_structural()

qa_report_content_parvar()

qa_report_content_covariates()
}
\value{
Latex/HTML for the section
}
\description{
These function generate the different content section for the QA report
}
